Add optional support for reset attestation command.
This commit adds a new command (0x55) and a new flag
to enable it (0x02) that will reset the attestation
private key and certificate without resetting the
chip key or counters. Once the command is issued and
accepted, the certificate will need to be loaded as
if it had never been loaded before.

If this command is issued via NFC, an artificial delay
of several seconds is introduced in order to reduce
the chances of unintentional denial of service.

By default this command is disabled. To enable it,
you must pass flag 0x02 in the parameters at applet
load.