Change iOS work-around to not break Windows. Fixes #2.
Apparently the iOS work-around, as implemented in commit 8b58c4cdcae,
has the side effect of causing Windows to reject the authenticator
entirely---see issue #2.

However, it was found that if we return `SW_INS_NOT_SUPPORTED` for
the FIDO2 NFC CTAP command (instead of `SW_CLA_NOT_SUPPORTED` or
mimicking a CTAP2 error response), we end up with a CTAP1 authenticator
that works on both iOS and Windows.

So this change reverts the previous work-around for iOS and makes
sure that we return `SW_INS_NOT_SUPPORTED` for any CLA with the highest
bit set.

Thanks go to Anurag Sharma (@annu0412) for discovering this issue.